requirements.txt
alembic==1.4.2
altgraph==0.17
Bootstrap-Flask==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudinary==1.21.1
cx-Freeze==6.1
cycler==0.10.0
dominate==2.5.1
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.2.2
numpy==1.19.0
pefile==2019.4.18
Pillow==7.2.0
pygame==1.9.6
PyInstaller==3.6
pyparsing==2.4.7
pyperclip==1.8.0
pytaglib==1.4.6
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pywin32-ctypes==0.2.0
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
supervisor==4.2.0
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
